+++ context.py
@@ -115,6 +115,9 @@ def get_settings():
     # Debug Mode
     settings["debug_mode"] = s.get("debug_mode", False)
 
+    # Content Types
+    settings["content_types"] = s.get("content_types", [])
+
     # Indicate whether switch project back to original 
     # after `deploy` or `retrieve from other server` is executed
     settings["switch_back_after_migration"] = s.get("switch_back_after_migration", True)